backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.20
pandas>=2.0.0
networkx>=3.0
openpyxl>=3.1.0
python-multipart>=0.0.5
requests>=2.31.0
